cpio_utils: fix compiler signedness warning
Fix gcc warning: comparison of integer expressions of different
signedness: 'size_t' {aka 'long unsigned int'} and 'int'.

Type-casted here instead of changed DecryptState's outlen type
since the OpenSSL functions ultimately called use int *outl.

+++ core/cpio_utils.c
@@ -161,7 +161,7 @@ static int decrypt_step(void *state, void *buffer, size_t size)
 	int inlen;
 
 	if (s->outlen != 0) {
-		if (size > s->outlen) {
+		if ((int)size > s->outlen) {
 			size = s->outlen;
 		}
 		memcpy(buffer, s->output, size);
@@ -198,7 +198,7 @@ static int decrypt_step(void *state, void *buffer, size_t size)
 	}
 
 	if (s->outlen != 0) {
-		if (size > s->outlen) {
+		if ((int)size > s->outlen) {
 			size = s->outlen;
 		}
 		memcpy(buffer, s->output, size);
